Fix `TlsGetValue` & more
`TlsGetValue` disambiguates 0 and an error by relying on `GetLastError`. Depending on the program state, `GetLastError` could be non-0, even though `TlsGetValue` succeeded. Resolve this by always calling `SetLastError`. This matches the behavior described by the documentation.

Additionally, when reading resources, later versions of mwcc and mwld call `GetModuleHandleA` with the program path, and then call `LoadStringA` on that handle. Support this behavior by _actually_ loading the PE at the path passed in to `GetModuleHandleA`, instead of assuming it's the current program. (This is useful because sjiswrap relies on overriding `GetModuleFileNameA`)

Other changes:
- Add ms-win-crt exit & run atexit funcs
- Implements vcruntime memmove