Add support for $PATH and Additional Path Utilities
Adds the ability to find the PE executable in $PATH. If not found, the
legacy behavior remains.

Adds the `--unix-path`/`-du` and `--dos-path`/`-ud` options to provide
the ability to convert between paths that might be created by utilities
run by wibo or local filesystem paths that will be passed to programs
run by wibo.